+++ docs-demo/basic/fixed-mode/FixedMode.vue
@@ -2,6 +2,8 @@
 import { ref } from 'vue';
 import StkTable from '../../StkTable.vue';
 import { StkTableColumn } from '../../../src/StkTable/index';
+import RangeInput from '../../components/RangeInput.vue';
+
 const width = ref(50);
 const columns: StkTableColumn<any>[] = [
     { dataIndex: 'id', title: 'id(100px)', width: '100px' },
@@ -15,7 +17,7 @@ const data = new Array(200).fill(0).map((it, index) => {
 
 <template>
     <div>
-        <input v-model="width" type="range" min="0" max="100" /> <span>width: {{ width }}%</span>
+        <RangeInput v-model="width" min="0" max="100" label="width" suffix="%"></RangeInput>
         <StkTable
             style="height: 150px"
             row-key="id"